Briscoe - Updated VisionSub and added VisionRangeCmd
In VisionSub, I added a getDistanceinInches function which finds any target's distance with a unit conversion of inches. The GetNumTargets has updated with its output parameters including two yaws which simplifies the GetYaw function code. The VisionRangeCmd is created and has the ability to have the robot get in range of a target with a deadzone. VisionSub remains not retested, and VisionRangeCmd is untested.